1.0
refactor: change project structure
