+++ src/components/generator/api.js
@@ -1,0 +1,8 @@
+const express = require('express')
+const controller = require('./controller')
+//
+const routes = express.Router()
+//
+routes.get('/generator', controller.create)
+//
+module.exports = routes
