+++ src/components/generator/model.js
@@ -1,0 +1,5 @@
+class Generator{
+    id//id do usuário
+    pass //senha gerada
+    create_time// datetime da criação
+}
