+++ src/components/generator/service.js
@@ -1,0 +1,9 @@
+const crypto = require('crypto')
+
+module.exports = {
+    create(){
+        var passwd = crypto.randomBytes(5).toString('hex')
+        //
+        return passwd
+    }
+}
