--- src/routes/auth.js
@@ -1,1 +1,0 @@
-
