+++ src/server.js
@@ -1,0 +1,5 @@
+const app = require('./app')
+
+app.listen(3333, () => {
+    console.log('Conectado')
+})
