--- src/services/knex.js
@@ -1,6 +1,0 @@
-const knex = require('knex')
-const conf = require('../../knexfile')
-
-const conn = knex(conf.development)
-
-module.exports = conn
